+++ src/Bridge/StorageFacade.php
@@ -31,6 +31,7 @@ use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeConfiguration\P
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeStorage\PortalNodeStorageClearActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeStorage\PortalNodeStorageDeleteActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeStorage\PortalNodeStorageGetActionInterface;
+use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeStorage\PortalNodeStorageListActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\PortalNodeStorage\PortalNodeStorageSetActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\ReceptionRouteListActionInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\Action\Route\RouteCreateActionInterface;
@@ -248,6 +249,11 @@ class StorageFacade extends AbstractSingletonStorageFacade
     protected function createPortalNodeStorageGetAction(): PortalNodeStorageGetActionInterface
     {
         return new PortalNodeStorageGet($this->getQueryFactory());
+    }
+
+    protected function createPortalNodeStorageListAction(): PortalNodeStorageListActionInterface
+    {
+        return new PortalNodeStorageList($this->getQueryFactory());
     }
 
     protected function createPortalNodeStorageSetAction(): PortalNodeStorageSetActionInterface